refactored to put html snippet site config to main module not gsmap module
