+++ django/main/models.py
@@ -1,0 +1,12 @@
+from solo.models import SingletonModel
+from django.contrib.gis.db import models
+
+class SiteConfiguration(SingletonModel):
+    search_enabled = models.BooleanField(default=True)
+    homepage_snippet = models.TextField(blank=True)
+
+    def __str__(self):
+        return "Site Configuration"
+
+    class Meta:
+        verbose_name = "Site Configuration"
